+++ assets/js/block/index.js
@@ -1,0 +1,13 @@
+import { registerBlockType } from '@wordpress/blocks';
+
+import './style.scss';
+
+/**
+ * Internal dependencies
+ */
+import Edit from './edit';
+import metadata from './block.json';
+
+registerBlockType( metadata.name, {
+	edit: Edit,
+} );